CREATED BY INAYAT JUPYTER NOTEBOOK
Movie Ticket Booking System for Avengers Phase Three (2016-2019)
This Python program allows users to book movie tickets for movies belonging to the Avengers Phase Three series, which includes popular titles released between 2016 and 2019. Users can select a movie from the provided list and input their age and the number of tickets they wish to book. If the user meets the age requirement for the selected movie and if there are enough tickets available, the program confirms the booking and updates the available ticket count accordingly. If the user is too young or if there are insufficient tickets available, appropriate messages are displayed.

Features:
User-friendly interface for selecting movies and booking tickets.
Age verification to ensure users meet the movie's age restriction.
Real-time update of available ticket count after each booking.
Option to end the booking process at any time.
Technologies Used:
Python
How to Use:
Run the Python script.
Enter the name of the movie you want to watch from the Avengers Phase Three series.
Input your age and the number of tickets you wish to book.
Follow the on-screen prompts to complete the booking process.
Optionally, choose to end the booking process if desired.